sugg(*): made the changes as per the suggestion
